Split language facts file
