+++ src/services/cssNavigation.ts
@@ -9,7 +9,7 @@ import * as nls from 'vscode-nls';
 import { DocumentContext } from '../cssLanguageTypes';
 import * as nodes from '../parser/cssNodes';
 import { Symbols } from '../parser/cssSymbolScope';
-import { getColorValue, hslFromColor } from '../services/languageFacts';
+import { getColorValue, hslFromColor } from './languageFacts';
 import { endsWith, startsWith } from '../utils/strings';
 
 const localize = nls.loadMessageBundle();
